Added before_upgrade test
Added before_upgrade test for the metric,
also added fixture for creating non-migratable vm.
