+++ tests/install_upgrade_operators/product_upgrade/test_upgrade_iuo.py
@@ -2,6 +2,7 @@ import logging
 
 import pytest
 
+from tests.install_upgrade_operators.constants import KUBEVIRT_VMI_NUMBER_OF_OUTDATED
 from tests.install_upgrade_operators.product_upgrade.utils import (
     process_alerts_fired_during_upgrade,
     verify_nodes_labels_after_upgrade,
@@ -13,6 +14,7 @@ from tests.upgrade_params import (
     IUO_CNV_ALERT_ORDERING_NODE_ID,
     IUO_UPGRADE_TEST_DEPENDENCY_NODE_ID,
 )
+from tests.utils import validate_metrics_value
 from utilities.constants import DEPENDENCY_SCOPE_SESSION
 from utilities.data_collector import collect_alerts_data
 
@@ -23,6 +25,16 @@ LOGGER = logging.getLogger(__name__)
 @pytest.mark.sno
 @pytest.mark.upgrade
 class TestUpgradeIUO:
+    """Pre-upgrade tests"""
+
+    @pytest.mark.polarion("CNV-11749")
+    def test_metric_kubevirt_vmi_number_of_outdated_before_upgrade(self, prometheus, cirros_vm_with_node_selector):
+        validate_metrics_value(
+            prometheus=prometheus,
+            metric_name=KUBEVIRT_VMI_NUMBER_OF_OUTDATED,
+            expected_value="0",
+        )
+
     """Post-upgrade tests"""
 
     @pytest.mark.polarion("CNV-9081")
@@ -85,5 +97,5 @@ class TestUpgradeIUO:
     def test_metric_kubevirt_vmi_number_of_outdated_after_upgrade(self, prometheus):
         wait_for_greater_than_zero_metric_value(
             prometheus=prometheus,
-            metric_name="kubevirt_vmi_number_of_outdated",
+            metric_name=KUBEVIRT_VMI_NUMBER_OF_OUTDATED,
         )
