+++ tests/observability/storage/test_hpp_observability.py
@@ -3,7 +3,7 @@ import logging
 import pytest
 
 from tests.observability.storage.constants import HPP_NOT_READY
-from tests.observability.utils import validate_metrics_value
+from tests.utils import validate_metrics_value
 from utilities.constants import (
     CRITICAL_STR,
     HOSTPATH_PROVISIONER_OPERATOR,
