+++ tests/observability/virt/test_virt_operator_leading.py
@@ -1,7 +1,7 @@
 import pytest
 
 from tests.observability.constants import KUBEVIRT_STR_LOWER
-from tests.observability.utils import validate_metrics_value
+from tests.utils import validate_metrics_value
 from utilities.constants import CRITICAL_STR
 from utilities.monitoring import validate_alerts
