+++ tests/storage/hpp/conftest.py
@@ -13,7 +13,7 @@ from tests.storage.hpp.utils import (
     update_node_taint,
     wait_for_desired_hpp_pods_running,
 )
-from tests.storage.utils import create_cirros_vm
+from tests.utils import create_cirros_vm
 from utilities.constants import TIMEOUT_5MIN
 from utilities.hco import add_labels_to_nodes
 from utilities.infra import (
