+++ tests/utils.py
@@ -5,13 +5,17 @@ import shlex
 import tarfile
 from contextlib import contextmanager
 from io import BytesIO
+from typing import Generator
 
 import bitmath
 import requests
 import xmltodict
 from bs4 import BeautifulSoup
+from kubernetes.dynamic import DynamicClient
 from kubernetes.dynamic.exceptions import ResourceNotFoundError
+from ocp_resources.datavolume import DataVolume
 from ocp_resources.kubevirt import KubeVirt
+from ocp_resources.prometheus import Prometheus
 from ocp_resources.resource import ResourceEditor
 from ocp_resources.template import Template
 from ocp_resources.virtual_machine import VirtualMachine
@@ -21,21 +25,33 @@ from timeout_sampler import TimeoutExpiredError, TimeoutSampler
 
 from utilities.constants import (
     DISK_SERIAL,
+    OS_FLAVOR_CIRROS,
     RHSM_SECRET_NAME,
     TIMEOUT_1SEC,
+    TIMEOUT_4MIN,
     TIMEOUT_5SEC,
     TIMEOUT_10MIN,
     TIMEOUT_10SEC,
+    TIMEOUT_15SEC,
     TIMEOUT_30MIN,
+    Images,
 )
 from utilities.hco import ResourceEditorValidateHCOReconcile, update_hco_annotations
-from utilities.infra import ExecCommandOnPod, get_artifactory_header
+from utilities.infra import (
+    ExecCommandOnPod,
+    get_artifactory_config_map,
+    get_artifactory_header,
+    get_artifactory_secret,
+    get_http_image_url,
+)
+from utilities.monitoring import get_metrics_value
 from utilities.virt import (
     VirtualMachineForTests,
     VirtualMachineForTestsFromTemplate,
     fedora_vm_body,
     get_created_migration_job,
     prepare_cloud_init_user_data,
+    running_vm,
     wait_for_migration_finished,
     wait_for_ssh_connectivity,
     wait_for_updated_kv_value,
@@ -545,3 +561,73 @@ def vm_object_from_template(
         network_multiqueue=param_dict.get("network_multiqueue"),
         ssh=param_dict.get("ssh", True),
     )
+
+
+@contextmanager
+def create_cirros_vm(
+    storage_class: str,
+    namespace: str,
+    client: DynamicClient,
+    dv_name: str,
+    vm_name: str,
+    node: str | None = None,
+    wait_running: bool = True,
+    volume_mode: str | None = None,
+    cpu_model: str | None = None,
+    annotations: dict[str, str] | None = None,
+) -> Generator[VirtualMachineForTests]:
+    artifactory_secret = get_artifactory_secret(namespace=namespace)
+    artifactory_config_map = get_artifactory_config_map(namespace=namespace)
+
+    dv = DataVolume(
+        name=dv_name,
+        namespace=namespace,
+        source="http",
+        url=get_http_image_url(image_directory=Images.Cirros.DIR, image_name=Images.Cirros.QCOW2_IMG),
+        storage_class=storage_class,
+        size=Images.Cirros.DEFAULT_DV_SIZE,
+        api_name="storage",
+        volume_mode=volume_mode,
+        secret=artifactory_secret,
+        cert_configmap=artifactory_config_map.name,
+    )
+    dv.to_dict()
+    dv_metadata = dv.res["metadata"]
+    with VirtualMachineForTests(
+        client=client,
+        name=vm_name,
+        namespace=dv_metadata["namespace"],
+        os_flavor=OS_FLAVOR_CIRROS,
+        memory_requests=Images.Cirros.DEFAULT_MEMORY_SIZE,
+        data_volume_template={"metadata": dv_metadata, "spec": dv.res["spec"]},
+        node_selector=node,
+        run_strategy=VirtualMachine.RunStrategy.ALWAYS,
+        cpu_model=cpu_model,
+        annotations=annotations,
+    ) as vm:
+        if wait_running:
+            running_vm(vm=vm, wait_for_interfaces=False)
+        yield vm
+
+
+def validate_metrics_value(
+    prometheus: Prometheus, metric_name: str, expected_value: str, timeout: int = TIMEOUT_4MIN
+) -> None:
+    samples = TimeoutSampler(
+        wait_timeout=timeout,
+        sleep=TIMEOUT_15SEC,
+        func=get_metrics_value,
+        prometheus=prometheus,
+        metrics_name=metric_name,
+    )
+    sample = None
+    try:
+        for sample in samples:
+            if sample:
+                LOGGER.info(f"metric: {metric_name} value is: {sample}, the expected value is {expected_value}")
+                if sample == expected_value:
+                    LOGGER.info("Metrics value matches the expected value!")
+                    return
+    except TimeoutExpiredError:
+        LOGGER.info(f"Metrics value: {sample}, expected: {expected_value}")
+        raise
